requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-admin-tools==0.9.0
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-import-export==2.3.0
django-js-asset==1.2.2
django-modeltranslation==0.15.2
django-mptt==0.11.0
django-redis==4.12.1
django-rest-multiple-models==2.1.3
django-taggit==1.3.0
djangorestframework==3.11.1
djangorestframework-recursive==0.1.2
drf-writable-nested==0.6.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
